+++ src/Api/BaseHttpApi.php
@@ -6,7 +6,7 @@ namespace Shoman4eg\Nalog\Api;
 use Http\Client\HttpClient;
 use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Message\ResponseInterface;
-use Shoman4eg\Nalog\Exception\Domain as DomainExceptions;
+use Shoman4eg\Nalog\ErrorHandler;
 use Shoman4eg\Nalog\Exception\DomainException;
 use Shoman4eg\Nalog\RequestBuilder;
 use Shoman4eg\Nalog\Util\JSON;
@@ -134,28 +134,15 @@ abstract class BaseHttpApi
      * Handle HTTP errors.
      *
      * Call is controlled by the specific API methods.
+     * Use ErrorHandler::handle instead of this
      *
      * @throws DomainException
+     *
+     * @deprecated
      */
     protected function handleErrors(ResponseInterface $response): void
     {
-        $body = (string)$response->getBody();
-        switch ($response->getStatusCode()) {
-            case 400:
-                throw new DomainExceptions\ValidationException($body);
-            case 401:
-                throw new DomainExceptions\UnauthorizedException();
-            case 403:
-                throw new DomainExceptions\ForbiddenException();
-            case 404:
-                throw new DomainExceptions\NotFoundException();
-            case 406:
-                throw new DomainExceptions\ClientException('Wrong Accept headers');
-            case 500:
-                throw new DomainExceptions\ServerException();
-            default:
-                throw new DomainExceptions\UnknownErrorException();
-        }
+        (new ErrorHandler())->handleResponse($response);
     }
 
     /**