+++ src/Exception/Domain/UnauthorizedException.php
@@ -8,4 +8,9 @@ namespace Shoman4eg\Nalog\Exception\Domain;
  */
 final class UnauthorizedException extends ClientException
 {
+    public function __construct($message)
+    {
+        $decodedMessage = json_decode($message, true);
+        parent::__construct($decodedMessage['message']);
+    }
 }